+++ block-demo-content.php
@@ -5,7 +5,7 @@
  * @Author: Timi Wahalahti
  * @Date:   2021-05-17 13:34:52
  * @Last Modified by:   Roni Laukkarinen
- * @Last Modified time: 2022-05-25 19:57:01
+ * @Last Modified time: 2022-05-26 15:18:55
  *
  * @package air-light
  */
@@ -27,13 +27,9 @@
 
             <p>This demo has custom blocks enabled, but air-light doesn't ship with them. You'll need to build your blocks yourself.</p>
 
-            <ul class="multiple-buttons-wrapper">
-              <a class="button js-trigger" href="https://airwptheme.com">
-                Features
-              </a>
-
-              <a class="button js-trigger" href="https://blocks.airwptheme.com">
-                Block library
+            <p class="button-wrapper">
+              <a class="button button-large" href="https://airwptheme.com">
+                See all features
               </a>
             </ul>
           </div>
@@ -44,7 +40,7 @@
 <div class="block-name-tag" aria-hidden="true"><p>hero</p></div>
 </section>
 
-<section class="block block-demo-centered">
+<section class="block block-demo-centered has-image">
   <div class="container">
     <h2 class="has-text-gradient">Gutenberg blocks make the website easy to update</h2>
     <p>Air-light has 100% Gutenberg support. Get magical with custom ACF + Gutenberg blocks.</p>
@@ -68,7 +64,7 @@
     <div class="col col-content">
       <div class="wrapper">
         <h2 class="has-text-gradient">Modular SCSS + CSS custom properties for everything</h2>
-        <p>We've got you covered with styles. Air-light is bloat-free! Just open _colors.scss and you are good to go. Did you know that Air-light production-ready CSS weights only 99KB without compression?</p>
+        <p>When it comes to Cascading Style Sheets, we've got you covered. Air-light is bloat-free! Just open _colors.scss and you are good to go. Did you know that Air-light production-ready CSS weights only 99KB without compression?</p>
         <p class="button-wrapper">
           <a class="button no-external-link-indicator has-icon" href="https://github.com/digitoimistodude/air-light">
             <svg width="16" height="16" viewBox="0 0 1792 1792" fill="currentColor" xmlns="http://www.w3.org/2000/svg" aria-hidden="true"><path d="M1664 896q0 251-146.5 451.5T1139 1625q-27 5-39.5-7t-12.5-30v-211q0-97-52-142 57-6 102.5-18t94-39 81-66.5 53-105T1386 856q0-121-79-206 37-91-8-204-28-9-81 11t-92 44l-38 24q-93-26-192-26t-192 26q-16-11-42.5-27T578 459.5 492 446q-44 113-7 204-79 85-79 206 0 85 20.5 150t52.5 105 80.5 67 94 39 102.5 18q-40 36-49 103-21 10-45 15t-57 5-65.5-21.5T484 1274q-19-32-48.5-52t-49.5-24l-20-3q-21 0-29 4.5t-5 11.5 9 14 13 12l7 5q22 10 43.5 38t31.5 51l10 23q13 38 44 61.5t67 30 69.5 7 55.5-3.5l23-4q0 38 .5 89t.5 54q0 18-13 30t-40 7q-232-77-378.5-277.5T128 896q0-209 103-385.5T510.5 231 896 128t385.5 103T1561 510.5 1664 896z"></path></svg>
@@ -85,7 +81,7 @@
           <li>&ndash;<span class="spacer"></span>&ndash;<span class="spacer"></span>color-brand-twitter: <span class="color" style="background-color: #1da1f2"></span> #1da1f2</li>
           <li>&ndash;<span class="spacer"></span>&ndash;<span class="spacer"></span>color-pattens-blue: <span class="color" style="background-color: #eaf1f8;"></span> #eaf1f8;</li>
           <li>&ndash;<span class="spacer"></span>&ndash;<span class="spacer"></span>color-science-blue: <span class="color" style="background-color: #0049d1;"></span> #0049d1;</li>
-          <li>&ndash;<span class="spacer"></span>&ndash;<span class="spacer"></span>color-scropion: <span class="color" style="background-color: #5f5f5f;"></span> #5f5f5f;</li>
+          <li>&ndash;<span class="spacer"></span>&ndash;<span class="spacer"></span>color-scorpion: <span class="color" style="background-color: #5f5f5f;"></span> #5f5f5f;</li>
           <li>&ndash;<span class="spacer"></span>&ndash;<span class="spacer"></span>color-valhalla: <span class="color" style="background-color: #2a2d3e;"></span> #2a2d3e;</li>
         </ul>
       </div>
@@ -189,7 +185,7 @@
     <h2 class="has-text-gradient">...and more. But still less.</h2>
     <p>Air-light light weights only under 20 KB gzipped and is accessible-ready, gutenberg-ready and multilingual-ready.</p>
     <p class="button-wrapper">
-      <a class="button no-external-link-indicator has-icon" href="https://docs.airwptheme.com/air-blocks/block-library">
+      <a class="button no-external-link-indicator has-icon" href="https://airwptheme.com">
         Check out all the features
       </a>
     </p>